Title: 최소 통과 시간, Time: 64ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240402/최소 통과 시간/minimum-transit-time.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 int n, m;
-vector<int> nums;
+vector<long long> nums;
 
 bool pass(long long time)
 {
